Add a date formatter (NSDateFormatter) and use it to display a more appropriate date description
We will create an instance property for an NSDateFormatter that will be created on-demand. This date formatter will be used to format the due date displayed in the todo list so that it's shorter and more appropriate.

Also, UITableViewCellStyleValue1 still looks too crowded, so let's try another cell style, UITableViewCellStyleSubtitle.